refactor: a huge refactor done - removed useless code, improved code, fixed bugs, and new menu style
The system has many areas for improvement. Therefore, some repeated and unnecessary code was
removed, and the responsibilities of each provider, component, and style were reviewed. This
improved maintainability, readability, and reusability. The authenticated data is now a state. Bugs,
such as the cellphone issue when completing a registration and async and await functions using
promise chaining were fixed. A new menu style was developed.